Update lambda for person tagging
