--- aws/src/main/python/cloudWatchReplicate/cloudWatchReplicate.yaml
@@ -1,20 +1,0 @@
-AWSTemplateFormatVersion: '2010-09-09'
-Transform: 'AWS::Serverless-2016-10-31'
-Description: An AWS Serverless Specification template describing your function.
-Resources:
-  cloudWatchReplicate:
-    Type: 'AWS::Serverless::Function'
-    Properties:
-      Handler: lambda_function.lambda_handler
-      Runtime: python3.7
-      CodeUri: .
-      Description: ''
-      MemorySize: 128
-      Timeout: 10
-      Role: 'arn:aws:iam::340032650202:role/BeamLambdaExecution'
-      Environment:
-        Variables:
-          AWS_ACCOUNT_ID: '340032650202'
-          EXCLUDED_INSTANCES: '[''i-0d7356ff7f3927e69'',''i-0b31829f99cc9a7ce'']'
-          SECONDS_TO_TRIGGER_IDLE_NOTIFICATION: '21600'
-          UPTIME_INTERVAL_TO_TRIGGER_NOTIFICATION_IN_DAYS: '5'
