--- aws/src/main/python/ec2InstanceChecker/ec2InstanceChecker.yaml
@@ -1,24 +1,0 @@
-AWSTemplateFormatVersion: '2010-09-09'
-Transform: 'AWS::Serverless-2016-10-31'
-Description: An AWS Serverless Specification template describing your function.
-Resources:
-  ec2InstanceChecker:
-    Type: 'AWS::Serverless::Function'
-    Properties:
-      Handler: lambda_function.lambda_handler
-      Runtime: python3.7
-      CodeUri: .
-      Description: ''
-      MemorySize: 128
-      Timeout: 3
-      Role: 'arn:aws:iam::340032650202:role/BeamLambdaExecution'
-      Environment:
-        Variables:
-          ENDPOINT: /api/chat.postMessage
-          EXCLUDED_INSTANCES: '[''i-0d7356ff7f3927e69'',''i-0b31829f99cc9a7ce'',''i-01dda95dee4c4714f'']'
-          SLACK_API_URL: slack.com
-          TOKEN: >-
-            <Replace with actual token>
-          USER_SLACK_IDS: >-
-            [{"rk_sharmaitATyahoo_com":"DHFV0MAN6"}, {"irishwithaxeATgmail_com":
-            "DMJ9E93EV"}]
